Clarify when QWeb host IDs are referenced
